Added scroll options to the table.
