+++ src/styles/DeploymentTable.css
@@ -4,11 +4,10 @@
   border-radius: 0.25rem; 
   padding: 1rem; 
   background-color: grey;
-  width: 100vw;
+  width: 100%;
   height: 100vh;
   display: flex;
   flex-direction: column;
   justify-content: flex-start;
   align-items: center;
-  margin-left: 20px;
 }
